Safari-609.1.11.1/LayoutTests/fast/dom/document-head.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This test checks to see if document.head is available, readonly, and the same element as what we would expect by getting it by other means.");

shouldBe("document.head", "document.getElementsByTagName('head')[0]");

document.head = 1;
shouldBeTrue("document.head !== 1");

document.documentElement.removeChild(document.head);
shouldBeNull("document.head");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
