third_party/blink/web_tests/fast/dom/document-head-expected.txt

This test checks to see if document.head is available, readonly, and the same element as what we would expect by getting it by other means.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.head is document.getElementsByTagName('head')[0]
PASS document.head !== 1 is true
PASS document.head is null
PASS successfullyParsed is true

TEST COMPLETE
